requirements.txt
annotated-types==0.7.0
anyio==4.4.0
boto3==1.35.4
botocore==1.35.4
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cohere==5.8.1
dependency-injector==4.41.0
exceptiongroup==1.2.2
fastapi==0.112.1
fastavro==1.9.5
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.6
idna==3.7
iniconfig==2.0.0
jmespath==1.0.1
joblib==1.4.2
numpy==2.0.1
packaging==24.1
parameterized==0.9.0
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
threadpoolctl==3.5.0
tokenizers==0.20.0
tomli==2.0.1
tqdm==4.66.5
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing_extensions==4.12.2
urllib3==1.26.19
uvicorn==0.30.6
